requirements.txt
apache-airflow-providers-postgres==5.11.0
pandas==2.1.4
astro-run-dag